more dogfooding issues: MultipleMembers weren’t working in some cases: 	We use MultipleMembers when we end up with a serialized value that consists of more than one type.  In Pygame this occurs w/ a module and a MissingModule object.  There were several things that weren’t working with this: 		Goto definition into the std lib – we are missing a Locations override 		Documentation, Description, Short description – more missing overrides 		Completions – we were unioning together the types, but never saving the newly created set.    Therefore we’d always return no type information for their members.
When we fail to infer a return type from a function we still need to include an overload, just w/o a return type.

Update AnalysisSaveTests to account for these issues.


Plus some engineering excellence work:
Work to fix our incremental builds, although we’re still not getting proper incremental builds across the board.  It’s better from the command line, VS still isn’t doing that will.  Looks like the remaining issues are related to the VS SDK project system.
Clean up some warnings

+++ Release/Product/Python/VsPyProf/PythonApi.cpp
@@ -137,7 +137,7 @@ bool VsPyProf::GetUserToken(PyFrameObject* frameObj, DWORD_PTR& func, DWORD_PTR&
 				_registeredModules[module] = moduleName;
 				
 				// make sure we only have valid path chars, vsperfreport doesn't like invalid chars
-				for(int i = 0; i<filenameStr.length(); i++) {
+				for(size_t i = 0; i<filenameStr.length(); i++) {
 					if(filenameStr[i] == '<') {
 						filenameStr[i] = '(';
 					}else if(filenameStr[i] == '>') {
